cherry-pick penghui checkmarx fix (#476)

+++ intel_extension_for_transformers/optimization/trainer.py
@@ -122,7 +122,6 @@ class BaseTrainer():
         self.enable_bf16 = False
         self.orchestrate_opt = False
         self.dynamic_config = None
-        
 
     @property
     def resuming_checkpoint(self):
@@ -1071,7 +1070,7 @@ class BaseTrainer():
         Return:
             :obj:`torch.Tensor`: The tensor with training loss on this batch.
         """
-    
+
         model.train()
         inputs = self._prepare_inputs(inputs)
         tr_loss_sum = 0.0
@@ -1105,7 +1104,6 @@ class BaseTrainer():
             compression_loss = self.compression_ctrl.loss()
             loss += compression_loss
 
- 
         loss = loss / (self.dynamic_config.num_sandwich + 2)
         tr_loss_sum += loss
 
@@ -1134,7 +1132,7 @@ class BaseTrainer():
                 loss = self.deepspeed.backward(loss)
             else:
                 loss.backward()
-       
+
         # inplace distillation
         for i in range(self.dynamic_config.num_sandwich + 1):
             ## prepare inputs for sub-models
@@ -1202,7 +1200,7 @@ class BaseTrainer():
             ## backward
 
             # pylint: disable=E0401
-            
+
             if version.parse(__version__) < version.parse("4.20"):
                 if self.use_amp:
                     self.scaler.scale(loss).backward()
@@ -1229,8 +1227,6 @@ class BaseTrainer():
 
         return tr_loss_sum.detach()
 
-
-
     # pylint: disable=E1101
     def compute_loss(self, model, inputs, return_outputs=False):  # pragma: no cover
         """
@@ -1244,7 +1240,7 @@ class BaseTrainer():
         teacher_logits = inputs.pop("teacher_logits") if "teacher_logits" in inputs else None
 
         outputs = model(**inputs)
-    
+
         if self.in_training and hasattr(self, "component") and \
            hasattr(self.component, "criterion"):
             qa_output_merger = lambda outputs: torch.vstack([
@@ -1792,7 +1788,8 @@ class BaseTrainer():
                 tensor.raw_data = fp16_data.tobytes()
                 tensor.data_type = TensorProto.BFLOAT16
         onnx.save(model, onnx_save_path)
-        os.remove(fp32_path)
+        if os.path.isfile(fp32_path):
+            os.remove(fp32_path)
 
         if verbose:
             info = "The ONNX Model is exported to path: {0}".format(onnx_save_path)
@@ -2010,7 +2007,7 @@ class BaseTrainer():
         logger.info("Weight type: {}.".format(weight_type))
         logger.info("Activation type: {}.".format(activation_type))
 
-        # Calibrate_method, min/max method as default. 
+        # Calibrate_method, min/max method as default.
         if 'minmax' in calibrate_method:
             calibrate_method = ortq.CalibrationMethod.MinMax
         elif 'percentile' in calibrate_method:
@@ -2107,7 +2104,9 @@ class BaseTrainer():
                         int8_model.graph.initializer.append(new_tensor)
                 onnx.save(int8_model, onnx_save_path)
 
-        os.remove(fp32_path)
+        if os.path.isfile(fp32_path):
+            os.remove(fp32_path)
+
         info = "The ONNX Model is exported to path: {0}".format(onnx_save_path)
         logger.info("*" * len(info))
         logger.info(info)
@@ -2160,7 +2159,7 @@ class BaseTrainer():
                         self.model.config.num_hidden_layers,
                         length_drop_ratio=self.dynamic_config.const_rate,
                     )
-        
+
         if lc is not None:
             # set the model with length config
             if self.model.config.model_type == "distilbert":
@@ -2175,7 +2174,6 @@ class BaseTrainer():
             bert.set_length_config(lc)
             bert.set_output_attentions(True)
 
-           
     def run_evolutionary_search(self):
         assert self.dynamic_config is not None, \
             """
